move geoJson builder to utils
